+++ release-notes.md
@@ -4,7 +4,7 @@
 Changes:
 - added 2 configurations to chose the viewer icons and the order in which they appear:
   'ViewersOrdering' and 'ViewersIcons'.  The default configuration is identical to the 0.4.1 behaviour.
- 
+- adapted orthanc-share API
 
 0.4.2 (2022-10-28)
 ==================